I added details to the section on thom transversality.
This is necessary because our section on morse theory is just the observation that I bolded, repeated many times.

Taylor's notes before mine appear to give another perspective on what the hessian is.  Perhaps it would be good to connect this with the discussion of hessians and its cubic analogue from the class on 2/12.